requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
Flask==2.3.2
Flask-HTTPAuth==4.8.0
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
